perf(web): improve app workflow build performance. (#26310)
